collector: fail on collector init error
At first we thought it would be a good idea to let other collectors work
if a given collector was failing to initialize, but I now think this is
actually not a good design and quite a poor ux. If some actions are
options in collectors, they can silence those errors and log messages
instead.

Signed-off-by: Antoine Tenart <atenart@redhat.com>