CI: check the coding style of bpf programs
The project, as stated in CONTRIBUTING have a strong preference for
the Linux kernel coding style.
This patch enforces it by adding automatic checks in the CI for bpf
sources.
The patch uses and import the current checkpatch.pl and its
requirement spelling.txt.

The current checkpatch.pl version is:

https://github.com/torvalds/linux/blob/86aa721820952b793a12fc6e5a01734186c0c238/scripts/checkpatch.pl

Signed-off-by: Paolo Valerio <pvalerio@redhat.com>