collector/ns: avoid sending unfiltered events
The ns collector so far requires either an sk_buff or a net_device as
a parameter to gather the ns-related info.

Unfortunately, as of today relying on net_device alone doesn't allow to
retrive the ns data based on the context.
Tracking is on its way to get extended making possible to capture info
from non-skb aware functions, but for the time being, turn the ns
collector into an skb-requiring only hook.

DEFINE_HOOK_RAW() plus tracking checks in the code are still used here
in place of a plain DEFINE_HOOK() as a contract that this will be
partially reverted, extended and adapted incorporating the upcoming
context tracking.

Signed-off-by: Paolo Valerio <pvalerio@redhat.com>